+++ rdbms/Conn.hpp
@@ -255,6 +255,16 @@ public:
    * @return the list of the names of the types in the database
    */
   std::list<std::string> getTypeNames();
+
+  /**
+   * Returns the view names of the database
+   *
+   * If the underlying database technologies does not support type informations
+   * this method simply returns an empty list.
+   *
+   * @return the list of the names of the views in the database
+   */
+  std::list<std::string> getViewNames();
   
   /**
    * Get a pointer to the connection wrapper implementation