+++ scheduler/rdbms/ArchiveJob.hpp
@@ -35,7 +35,7 @@ class ArchiveJob : public SchedulerDatabase::ArchiveJob {
  public:
 
    ArchiveJob();
-   ArchiveJob(bool jobOwned, uint64_t jobID, uint64_t mountID, std::string_view tapePool);
+   ArchiveJob(bool jobOwned, uint64_t jid, uint64_t mountID, std::string_view tapePool);
 
    void failTransfer(const std::string & failureReason, log::LogContext & lc) override;
 
@@ -45,7 +45,6 @@ class ArchiveJob : public SchedulerDatabase::ArchiveJob {
 
    bool m_jobOwned = false;
    uint64_t m_mountId = 0;
-   uint64_t m_jobId = 0;
    std::string m_tapePool;
 
 };