+++ scheduler/rdbms/postgres/ArchiveJobSummary.hpp
@@ -19,6 +19,7 @@
 
 #include "scheduler/rdbms/postgres/Enums.hpp"
 #include "rdbms/NullDbValue.hpp"
+#include <sstream>
 
 namespace cta::schedulerdb::postgres {
 
@@ -32,6 +33,8 @@ struct ArchiveJobSummaryRow {
   time_t oldestJobStartTime = std::numeric_limits<time_t>::max();
   uint16_t archivePriority = 0;
   uint32_t archiveMinRequestAge = 0;
+  uint32_t lastUpdateTime = 0;
+  uint32_t lastJobUpdateTime = 0;
 
   ArchiveJobSummaryRow() = default;
 
@@ -55,6 +58,8 @@ struct ArchiveJobSummaryRow {
     oldestJobStartTime   = rset.columnUint64("OLDEST_JOB_START_TIME");
     archivePriority      = rset.columnUint16("ARCHIVE_PRIORITY");
     archiveMinRequestAge = rset.columnUint32("ARCHIVE_MIN_REQUEST_AGE");
+    lastUpdateTime     = rset.columnUint32("LAST_UPDATE_TIME");
+    lastJobUpdateTime  = rset.columnUint32("LAST_JOB_UPDATE_TIME");
     return *this;
   }
 
@@ -68,6 +73,8 @@ struct ArchiveJobSummaryRow {
     params.add("oldestJobStartTime", oldestJobStartTime);
     params.add("archivePriority", archivePriority);
     params.add("archiveMinRequestAge", archiveMinRequestAge);
+    params.add("lastUpdateTime", lastUpdateTime);
+    params.add("lastJobUpdateTime", lastJobUpdateTime);
   }
 
   /**
@@ -76,6 +83,11 @@ struct ArchiveJobSummaryRow {
    * @return result set containing all rows in the table
    */
   static rdbms::Rset selectNotOwned(Transaction &txn) {
+    // locking the view until commit (DB lock released)
+    // this is to prevent tape servers counting the rows all at the same time
+    const char *const lock_sql = "LOCK TABLE ARCHIVE_JOB_SUMMARY IN ACCESS EXCLUSIVE MODE";
+    auto stmt = txn.conn().createStmt(lock_sql);
+    stmt.executeNonQuery();
     const char *const sql = "SELECT "
       "MOUNT_ID,"
       "STATUS,"
@@ -85,13 +97,16 @@ struct ArchiveJobSummaryRow {
       "JOBS_TOTAL_SIZE,"
       "OLDEST_JOB_START_TIME,"
       "ARCHIVE_PRIORITY,"
-      "ARCHIVE_MIN_REQUEST_AGE "
+      "ARCHIVE_MIN_REQUEST_AGE, "
+      "LAST_JOB_UPDATE_TIME, "
+      "LAST_UPDATE_TIME "
     "FROM ARCHIVE_JOB_SUMMARY WHERE "
     "MOUNT_ID IS NULL";
 
-    auto stmt = txn.conn().createStmt(sql);
+    stmt = txn.conn().createStmt(sql);
     return stmt.executeQuery();
   }
+
 };
 
 } // namespace cta::schedulerdb::postgres